ZA: Move the South African home view in views.py to its own file
