+++ pombola/south_africa/views/__init__.py
@@ -1,1 +1,2 @@
+from .home import *
 from .views import *
